# perf: improve performance of the sign expr module's methods