--- a/crates/proof-of-sql/src/base/bit/bit_mask_utils.rs
+++ b/crates/proof-of-sql/src/base/bit/bit_mask_utils.rs
@@ -2,6 +2,7 @@ use crate::base::scalar::ScalarExt;
 use bnum::types::U256;
 use core::ops::Shl;
 
+#[inline]
 pub fn make_bit_mask<S: ScalarExt>(x: S) -> U256 {
     let x_as_u256 = x.into_u256_wrapping();
     if x > S::MAX_SIGNED {
@@ -13,6 +14,7 @@ pub fn make_bit_mask<S: ScalarExt>(x: S) -> U256 {
     }
 }
 
+#[inline]
 pub fn is_bit_mask_negative_representation(bit_mask: U256) -> bool {
     bit_mask & (U256::ONE.shl(255)) == U256::ZERO
 }